Fix bug: Tooltip not appearing for certain points (#501)
This fixes an issue where the tooltip would not appear for certain data
points. I had thirty data points in a line chart, and the tooltip was
not working for points 2-9. Looking further, I noticed that the array
of points along the x axis was not being sorted in the correct order.
The numbers were being sorted in alphabetical order (which explains why
10 came after 1), so I added a compare function in the related sorts to
sort the number in xSet numerically.